mz: remove init.c, merge with mausezahn.c
